Moved index ts to src folder
